Social_Distancing_with_Bird_eye/requirements.txt
streamlit
opencv-python-headless
clint
cmake
streamlit_drawable_canvas
requests
Pillow